플레이어와 적 동일기능 상속 스크립트 추가
CharacterBase 추가

+++ Assets/Scripts/Default/ICombat.cs
@@ -11,11 +11,11 @@ public class ICombat
 
     public interface IDamage
     {
-        void TakeDamage(int damage);
+        void TakeDamage(float damage);
     }
     public interface IHealth
     {
-        int CurrentHealth { get; }
+        float CurrentHealth { get; }
         int MaxHealth { get; }
         void Die();
     }
